+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5] - 2019-06-16
+
+### Changed
+- Convert installed module directories automatically to Pascal Case.
+- Modify base path when installing from the site directory to avoid nested site/site/ directories.
+
 ## [0.0.4] - 2019-06-10
 
 ### Fixed